qcm controller

+++ src/main/java/org/example/projet_tuto/Repository/QCMRespository.java
@@ -3,5 +3,5 @@ package org.example.projet_tuto.Repository;
 import org.example.projet_tuto.entities.QCM;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QCMRespository extends JpaRepository<QCM, Integer> {
+public interface QCMRespository extends JpaRepository<QCM, Long> {
 }
